+++ examples/notes/shared/src/lib.rs
@@ -1,20 +1,10 @@
 #![allow(clippy::unsafe_derive_deserialize)]
 pub mod app;
 pub mod capabilities;
-#[cfg(any(feature = "wasm_bindgen", feature = "uniffi"))]
 mod ffi;
 
 pub use app::*;
 pub use crux_core::Core;
 pub use crux_kv as kv;
 
-#[cfg(any(feature = "wasm_bindgen", feature = "uniffi"))]
 pub use ffi::CoreFFI;
-
-#[cfg(feature = "uniffi")]
-const _: () = assert!(
-    uniffi::check_compatible_version("0.29.4"),
-    "please use uniffi v0.29.4"
-);
-#[cfg(feature = "uniffi")]
-uniffi::setup_scaffolding!();